requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
black==23.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
cleantext==1.1.4
click==8.1.3
cryptography==40.0.2
exceptiongroup==1.1.1
fastapi==0.95.2
fire==0.5.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
joblib==1.3.2
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
openai==0.28.1
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.5.1
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==1.10.8
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
termcolor==2.3.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.6.2
urllib3==2.0.7
uvicorn==0.22.0
yarl==1.9.2
